+++ arch/zsyscalls.go
@@ -19,7 +19,7 @@
 
 package arch
 
-// Based on Linux v6.14.
+// Based on Linux v6.15.
 
 var syscallsARM = map[int]string{
 	0:      "restart_syscall",
@@ -441,6 +441,7 @@ var syscallsARM = map[int]string{
 	464:    "getxattrat",
 	465:    "listxattrat",
 	466:    "removexattrat",
+	467:    "open_tree_attr",
 	983041: "breakpoint",
 	983042: "cacheflush",
 	983043: "usr26",
@@ -793,6 +794,7 @@ var syscallsAARCH64 = map[int]string{
 	464: "getxattrat",
 	465: "listxattrat",
 	466: "removexattrat",
+	467: "open_tree_attr",
 }
 
 var syscalls386 = map[int]string{
@@ -1252,6 +1254,7 @@ var syscalls386 = map[int]string{
 	464: "getxattrat",
 	465: "listxattrat",
 	466: "removexattrat",
+	467: "open_tree_attr",
 }
 
 var syscallsX32 = map[int]string{
@@ -1634,6 +1637,7 @@ var syscallsX32 = map[int]string{
 	464: "getxattrat",
 	465: "listxattrat",
 	466: "removexattrat",
+	467: "open_tree_attr",
 	512: "rt_sigaction",
 	513: "rt_sigreturn",
 	514: "ioctl",
@@ -2052,4 +2056,5 @@ var syscallsX86_64 = map[int]string{
 	464: "getxattrat",
 	465: "listxattrat",
 	466: "removexattrat",
+	467: "open_tree_attr",
 }